api/requirements.txt
Flask
Flask-SQLAlchemy
google-cloud-aiplatform
python-dotenv
medspacy
